refactor(buffett-indicator): split the model pipeline and clarify series provenance
Break the god-module into focused seams, load Yahoo as DatedSeries instead of fake FRED, and draw parity via referenceLines while keeping zone-colored history.
