+++ src/layout-marketplace/pane-imagery.ts
@@ -22,7 +22,7 @@ const IMAGERY_RULES: ReadonlyArray<readonly [string, PaneImagery]> = [
   ["sectors", "heatmap"],
   ["treemap", "heatmap"],
   ["fear-greed", "gauge"],
-  ["buffett", "gauge"],
+  ["buffett", "chart"],
   ["kelly", "gauge"],
   ["gauge", "gauge"],
   ["calendar", "calendar"],
